refactor(tools): decompose the coded-action validator into cohesive modules
One 1221-line file with nine gates read as a service rather than a utility. The
CLI stays at the same path, because the invocation string is prescribed in three
prose files and the test suite path-loads it; everything behind it moves into
tools/coded_action/.

The split follows what each part is responsible for, not what it is called:

  turtle.py        lexing an action TTL by masking and depth-scanning text
  action_model.py  the domain model those tokens build, and the func: marker
  job_source.py    facts read out of TypeScript the same way
  contract.py      the bridge to the deriver: lower a contract, or name why not
  pairs.py         find the (TTL, job) pairs and the schema they resolve against
  typecheck.py     compile one job against a stub of the SDK
  verdict.py       gate bookkeeping and the JSON shape, in one place
  gates.py         the nine predicates
  __init__.py      why any of this exists; nothing importable

The entry point keeps argparse, the gate order and the exit code, and nothing
else. Actions get CLIs; predicates get functions. Nine gate CLIs would have
pushed aggregation into skill prose and re-read the TTL, the .ofn, the job and
one contract derivation once per gate.

Two things this fixes on the way. contract.load_deriver now walks parents for
tools/entry_points.py, matching what the deploy skill's staging step does; the
two previously disagreed about where tools/ is. And the entry point's docstring
no longer claims a type<T>() contract "passes with a note because the SDK
derivation supplies the flag" -- the gate proves it by running the deriver.

entry_points.py deliberately does not move and gains no sibling imports: it is
path-loaded from two different trees, and a module loaded that way has no
package context, so a sibling import would break it.

Verified byte-identical against the committed golden after every module moved,
and under all three invocation modes: as a script from the repo root, as a
script from elsewhere, and path-loaded in-process. 30 tests pass.
